Major rewrite Put all IO arount usage examples into the Lexemes class Simplify the functions in usage_examples.py Move classes from misc.py to own files. Remove unused json_cache.py Comment out all dead code.
